PowerBI_COE/Auto Licence Cleaning/auto_clean_licences/requirements.txt
google-cloud-bigquery==3.25.0
requests==2.32.3

# Dev / tests uniquement
pytest==8.3.4
